server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
Pillow==10.1.0
requests==2.31.0
